Changes to be committed: 	modified:   src/api/java/com/liefeng/intf/property/ISysSecurityService.java 	modified:   src/api/java/com/liefeng/property/vo/sys/SysMenuVo.java 	modified:   src/api/java/com/liefeng/property/vo/sys/SysRoleMenuVo.java 	modified:   src/api/java/com/liefeng/property/vo/sys/SysRoleUserVo.java 	modified:   src/main/java/com/liefeng/property/domain/sys/SysMenuContext.java 	modified:   src/main/java/com/liefeng/property/domain/sys/SysRoleContext.java 	modified:   src/main/java/com/liefeng/property/domain/sys/SysRoleMenuContext.java 	modified:   src/main/java/com/liefeng/property/po/sys/SysMenuPo.java 	modified:   src/main/java/com/liefeng/property/po/sys/SysRoleMenuPo.java 	modified:   src/main/java/com/liefeng/property/po/sys/SysRoleUserPo.java 	modified:   src/main/java/com/liefeng/property/repository/sys/SysMenuRepository.java 	modified:   src/main/java/com/liefeng/property/repository/sys/SysRoleMenuRepository.java 	modified:   src/main/java/com/liefeng/property/service/SysSecurityService.java 	modified:   src/test/java/com/liefeng/property/test/integration/security/SecurityTest.java

+++ src/api/java/com/liefeng/property/vo/sys/SysRoleMenuVo.java
@@ -16,9 +16,9 @@ public class SysRoleMenuVo extends BaseVoValue{
 
 	private Long id;
 	
-	private Integer roleId;
+	private Long roleId;
 
-	private Integer menuId;
+	private Long menuId;
 
 	public Long getId() {
 		return id;
@@ -28,19 +28,19 @@ public class SysRoleMenuVo extends BaseVoValue{
 		this.id = id;
 	}
 
-	public Integer getRoleId() {
+	public Long getRoleId() {
 		return roleId;
 	}
 
-	public void setRoleId(Integer roleId) {
+	public void setRoleId(Long roleId) {
 		this.roleId = roleId;
 	}
 
-	public Integer getMenuId() {
+	public Long getMenuId() {
 		return menuId;
 	}
 
-	public void setMenuId(Integer menuId) {
+	public void setMenuId(Long menuId) {
 		this.menuId = menuId;
 	}
 }
