+++ src/main/java/com/liefeng/property/po/sys/SysRoleMenuPo.java
@@ -28,10 +28,10 @@ public class SysRoleMenuPo extends BaseValue{
 	private Long id;
 	
 	@Column(name = "role_id")
-	private Integer roleId;
+	private Long roleId;
 	
 	@Column(name = "menu_id")
-	private Integer menuId;
+	private Long menuId;
 	
 	@Column(name = "oem_code")
 	private String oemCode;
@@ -44,19 +44,19 @@ public class SysRoleMenuPo extends BaseValue{
 		this.id = id;
 	}
 
-	public Integer getRoleId() {
+	public Long getRoleId() {
 		return roleId;
 	}
 
-	public void setRoleId(Integer roleId) {
+	public void setRoleId(Long roleId) {
 		this.roleId = roleId;
 	}
 
-	public Integer getMenuId() {
+	public Long getMenuId() {
 		return menuId;
 	}
 
-	public void setMenuId(Integer menuId) {
+	public void setMenuId(Long menuId) {
 		this.menuId = menuId;
 	}
