+++ src/main/java/com/liefeng/property/po/sys/SysRoleUserPo.java
@@ -28,10 +28,10 @@ public class SysRoleUserPo extends BaseValue{
 	private Long id;
 	
 	@Column(name = "role_id")
-	private Integer roleId;
+	private Long roleId;
 	
 	@Column(name = "user_id")
-	private Integer userId;
+	private Long userId;
 	
 	@Column(name = "oem_code")
 	private String oemCode;
@@ -44,19 +44,19 @@ public class SysRoleUserPo extends BaseValue{
 		this.id = id;
 	}
 
-	public Integer getRoleId() {
+	public Long getRoleId() {
 		return roleId;
 	}
 
-	public void setRoleId(Integer roleId) {
+	public void setRoleId(Long roleId) {
 		this.roleId = roleId;
 	}
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
